+++ src/bindings/c++/src/write_features-c++.h
@@ -61,9 +61,9 @@ public:
     int properties() const;
     const std::vector<SailCompression>& compressions() const;
     SailCompression default_compression() const;
-    int compression_level_min() const;
-    int compression_level_max() const;
-    int compression_level_default() const;
+    double compression_level_min() const;
+    double compression_level_max() const;
+    double compression_level_default() const;
 
     sail_status_t to_write_options(write_options *swrite_options) const;
 
@@ -80,9 +80,9 @@ private:
     write_features& with_properties(int properties);
     write_features& with_compressions(const std::vector<SailCompression> &compressions);
     write_features& with_default_compression(SailCompression default_compression);
-    write_features& with_compression_level_min(int compression_level_min);
-    write_features& with_compression_level_max(int compression_level_max);
-    write_features& with_compression_level_default(int compression_level_default);
+    write_features& with_compression_level_min(double compression_level_min);
+    write_features& with_compression_level_max(double compression_level_max);
+    write_features& with_compression_level_default(double compression_level_default);
 
     const sail_write_features* sail_write_features_c() const;
